Reorganize components some more
