+++ assets/src/components/with-edit-featured-image.js
@@ -6,7 +6,8 @@ import { dispatch, withSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { hasMinimumDimensions, getMinimumFeaturedImageDimensions, getMinimumStoryPosterDimensions } from '../stories-editor/helpers';
+import { getMinimumStoryPosterDimensions } from '../stories-editor/helpers';
+import { hasMinimumDimensions, getMinimumFeaturedImageDimensions } from '../common/helpers';
 
 /**
  * Gets a wrapped version of a block's edit component that conditionally sets the featured image (only for AMP Story posts).
@@ -19,15 +20,20 @@ import { hasMinimumDimensions, getMinimumFeaturedImageDimensions, getMinimumStor
  */
 export default ( BlockEdit ) => {
 	return withSelect( ( select, ownProps ) => {
-		const featuredImage = select( 'core/editor' ).getEditedPostAttribute( 'featured_media' );
-		const isCorrectBlock = ( 'core/image' === ownProps.name || 'amp/amp-story-page' === ownProps.name );
+		const { getMedia } = select( 'core' );
+		const { getEditedPostAttribute } = select( 'core/editor' );
+		const { getSelectedBlock } = select( 'core/block-editor' );
+		const { editPost } = dispatch( 'core/editor' );
 
-		if ( featuredImage || ! isCorrectBlock || ! ownProps.attributes ) {
+		const featuredImage = getEditedPostAttribute( 'featured_media' );
+		const isRelevantBlock = ( 'core/image' === ownProps.name || 'amp/amp-story-page' === ownProps.name );
+
+		if ( featuredImage || ! isRelevantBlock || ! ownProps.attributes ) {
 			return;
 		}
 
 		const selectedMediaId = ownProps.attributes.mediaId || ownProps.attributes.id;
-		const selectedBlock = select( 'core/block-editor' ).getSelectedBlock();
+		const selectedBlock = getSelectedBlock();
 		if ( ! selectedMediaId || ! selectedBlock || ! selectedBlock.attributes ) {
 			return;
 		}
@@ -39,9 +45,13 @@ export default ( BlockEdit ) => {
 		}
 
 		// Conditionally set the selected image as the featured image.
-		const media = select( 'core' ).getMedia( selectedMediaId );
-		if ( media && media.media_details && hasMinimumDimensions( media.media_details, getMinimumFeaturedImageDimensions() ) && hasMinimumDimensions( media.media_details, getMinimumStoryPosterDimensions() ) ) {
-			dispatch( 'core/editor' ).editPost( { featured_media: selectedMediaId } );
+		const media = getMedia( selectedMediaId );
+		if (
+			media && media.media_details &&
+			hasMinimumDimensions( media.media_details, getMinimumFeaturedImageDimensions() ) &&
+			hasMinimumDimensions( media.media_details, getMinimumStoryPosterDimensions() )
+		) {
+			editPost( { featured_media: selectedMediaId } );
 		}
 	} )( ( props ) => {
 		return (
